Fix some bugs and added functionality
The user dont have to press the equal-sign now to get a result
(operators work too).
If the user presses the equal-sign, it shows up in the calculation-
display.
Edited Display style, changed displays from div to input to scroll
while the user is typing.
Added some comments in the css-file for a better readability.